work in progress

+++ web/src/main/java/com/infoshare/eventmanagers/servlets/FavoriteManageServlet.java
@@ -19,23 +19,18 @@ public class FavoriteManageServlet extends HttpServlet {
 
     @Inject
     FavoriteService favoriteService;
-    @Inject
-    EventService eventService;
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         int favoriteId = Integer.parseInt(req.getParameter("favoriteId"));
         int userId = Integer.parseInt(req.getParameter("userId"));
-        favoriteService.addFavorite(userId,favoriteId);
+        String action = req.getParameter("action");
+        if (action.equals("remove")){
+            favoriteService.removeFavorite(userId,favoriteId);
+        }
+        else if(action.equals("add")){
+            favoriteService.addFavorite(userId,favoriteId);
+        }
+
     }
-
-    @Override
-    protected void doDelete(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        int favoriteId = Integer.parseInt(req.getParameter("favoriteId"));
-        int userId = Integer.parseInt(req.getParameter("userId"));
-        favoriteService.removeFavorite(userId,favoriteId);
-    }
-
-
-
 }
